move transition logic to application route
